+++ plugins/api/src/index.js
@@ -7,23 +7,24 @@ const randomCategory = require('./category/getRandom')
 
 const addMethod = function (models) {
   // doc methods
-  models.Doc.prototype.redirects = function () {
+  models.Doc.prototype.getRedirects = function () {
     return getRedirects(this, models.http)
   }
-  models.Doc.prototype.incomingLinks = function () {
+  models.Doc.prototype.getIncoming = function () {
     return getIncoming(this, models.http)
   }
-  models.Doc.prototype.pageViews = function () {
+  models.Doc.prototype.getPageViews = function () {
     return getPageViews(this, models.http)
   }
+
   // constructor methods
-  models.wtf.randomCategory = function (options) {
+  models.wtf.getRandomCategory = function (options) {
     return randomCategory(options, models.http)
   }
-  models.wtf.getTemplate = function (template, options) {
+  models.wtf.getTemplatePages = function (template, options) {
     return getTransclusions(template, options, models.http)
   }
-  models.wtf.getCategory = function (category, options) {
+  models.wtf.getCategoryPages = function (category, options) {
     return getCategory(category, options, models.wtf)
   }
 }