+++ changelog.md
@@ -1,6 +1,7 @@
 <!--
 #### [unreleased]
 - **[breaking]** - move .random() to plugin-api
+- **[breaking]** - move .category() to plugin-api
 - deprecate plugin-category (moved to plugin-api)
 -->
 #### 8.5.1